more menu isolation, new multiplier numbers 831 lines now! 3k lines gone!!!!! woo hoo
